APERTA-7082 Fix problem when processing field.
Complex fields are delimited by `w:fldChar` elements with
`w:fldCharType` attributes with the value `begin`, `separate` or `end`.

After finding the end element, the sibling elements still need to be
processed, but we want, it seems, to process the parents in order to get
the right stylesheets, not the captured groups.